requirements.txt

streamlit
yfinance
numpy
plotly
scikit-learn
textblob
openai
python-dotenv
requests
beautifulsoup4
